create getElements file
