+++ packages/frontend/src/scripts/section-navigation/getElements.ts
@@ -1,0 +1,62 @@
+import { DESKTOP_PROJECT_NAVIGATION_IDS } from '../../components/project/navigation/DesktopProjectNavigation'
+import { MOBILE_PROJECT_NAVIGATION_IDS } from '../../components/project/navigation/MobileProjectNavigation'
+
+export function getDesktopElements() {
+  const container = document.querySelector(
+    `#${DESKTOP_PROJECT_NAVIGATION_IDS.container}`,
+  )
+  const list = container?.querySelector(
+    `#${DESKTOP_PROJECT_NAVIGATION_IDS.list}`,
+  )
+  const listHeader = container?.querySelector(
+    `#${DESKTOP_PROJECT_NAVIGATION_IDS.listHeader}`,
+  )
+  const summaryItem = list?.querySelector<HTMLAnchorElement>(
+    `a#${DESKTOP_PROJECT_NAVIGATION_IDS.summaryItem}`,
+  )
+  const sections = document.querySelectorAll('section')
+
+  if (!container || !list || !listHeader || !summaryItem) {
+    throw new Error('Could not find all elements for desktop navigation')
+  }
+
+  return {
+    container,
+    list,
+    listHeader,
+    summaryItem,
+    sections,
+  }
+}
+
+export function getMobileElements() {
+  const container = document.querySelector(
+    `#${MOBILE_PROJECT_NAVIGATION_IDS.container}`,
+  )
+  const list = container?.querySelector(
+    `#${MOBILE_PROJECT_NAVIGATION_IDS.list}`,
+  )
+  const summaryItem = container?.querySelector<HTMLAnchorElement>(
+    `a#${MOBILE_PROJECT_NAVIGATION_IDS.summaryItem}`,
+  )
+  const arrowLeft = container?.querySelector(
+    `#${MOBILE_PROJECT_NAVIGATION_IDS.arrowLeft}`,
+  )
+  const arrowRight = container?.querySelector(
+    `#${MOBILE_PROJECT_NAVIGATION_IDS.arrowRight}`,
+  )
+
+  const sections = document.querySelectorAll('section')
+
+  if (!container || !list || !summaryItem || !arrowLeft || !arrowRight) {
+    throw new Error('Could not find all elements for mobile navigation')
+  }
+
+  return {
+    list,
+    summaryItem,
+    arrowLeft,
+    arrowRight,
+    sections,
+  }
+}
